Comparable pairs in a partial order

Calculates the fraction of comparable pairs in a partial order.

Usage

comparable_pairs(P)

Arguments

P

A partial order as matrix object, e.g. calculated with neighborhood_inclusion or positional_dominance.

Value

Fraction of comparable pairs in P.

See also

Author

David Schoch

Examples

library(igraph)
g <- sample_gnp(100, 0.1)
P <- neighborhood_inclusion(g)
comparable_pairs(P)
#> [1] 0
# All pairs of vertices are comparable in a threshold graph
tg <- threshold_graph(100, 0.3)
P <- neighborhood_inclusion(g)
comparable_pairs(P)
#> [1] 0