Hyperbolic (centrality) index

The hyperbolic index is an index that considers all closed walks of even or odd length on induced neighborhoods of a vertex.

hyperbolic_index(g, type = "odd")

Arguments

g
igraph object.

type
string. 'even' if only even length walks should be considered. 'odd' (Default) if only odd length walks should be used.

Value

A vector containing centrality scores.

Details

The hyperbolic index is an illustrative index that should not be used for any serious analysis. Its purpose is to show that with enough mathematical trickery, any desired result can be obtained when centrality indices are used.

Examples

library(igraph)

g <- graph.empty(n=11,directed = FALSE)
g <- add_edges(g,c(1,11,2,4,3,5,3,11,4,8,5,9,5,11,6,7,6,8,
                   6,10,6,11,7,9,7,10,7,11,8,9,8,10,9,10))
hyperbolic_index(g,type = "odd")

#>  [1] 0.000e+00 0.000e+00 1.276e+00 4.219e-15 9.186e-01 3.555e+00 3.555e+00
#>  [8] 1.926e+00 1.926e+00 5.996e+00 1.501e+00

hyperbolic_index(g,type = "even")

#>  [1]  3.086  3.086  6.848  3.571  6.228 11.665 11.665  8.585  8.585 16.292
#> [11]  7.755