Check preservation

Checks if a partial ranking is preserved in the ranking induced by scores.

is_preserved(P, scores)

Arguments

P
A partial ranking as matrix object calculated with neighborhood_inclusion or positional_dominance.

scores
Numeric vector containing the scores of a centrality index.

Value

Logical scaler whether scores preserves the relations in P.

Details

In order for a score vector to preserve a partial ranking, the following condition must be fulfilled: P[u,v]==1 & scores[i]<=scores[j].

Examples

library(igraph)
# standard measures of centrality preserve the neighborhood inclusion preorder
g <- graph.empty(n=11,directed = FALSE)
g <- add_edges(g,c(1,11,2,4,3,5,3,11,4,8,5,9,5,11,6,7,6,8,
                   6,10,6,11,7,9,7,10,7,11,8,9,8,10,9,10))
P<-neighborhood_inclusion(g)

is_preserved(P,degree(g))

#> [1] TRUE

is_preserved(P,betweenness(g))

#> [1] TRUE

is_preserved(P,closeness(g))

#> [1] TRUE