Plot rank intervals

This function is deprecated. Use plot(rank_intervals(P)) instead

Usage

plot_rank_intervals(P, cent.df = NULL, ties.method = "min")

Arguments

P

A partial ranking as matrix object calculated with neighborhood_inclusion or positional_dominance.

cent.df

A data frame containing centrality scores of indices (optional). See Details.

ties.method

String specifying how ties are treated in the base rank function.

See also

Author

David Schoch

Examples

library(igraph)
data("dbces11")
P <- neighborhood_inclusion(dbces11)
if (FALSE) { # \dontrun{
plot_rank_intervals(P)
} # }

# adding index based rankings
cent_scores <- data.frame(
    degree = degree(dbces11),
    betweenness = round(betweenness(dbces11), 4),
    closeness = round(closeness(dbces11), 4),
    eigenvector = round(eigen_centrality(dbces11)$vector, 4)
)
if (FALSE) { # \dontrun{
plot_rank_intervals(P, cent.df = cent_scores)
} # }