Indirect relations in a network

Derive indirect relations for a given network. Observed relations, like presents or absence of a relation, are commonly not the center of analysis, but are transformed in a new set of indirect relation like geodesic distances among nodes. These transformations are usually an implicit step when centrality indices are used. Making this step explicit gives more possibilities, for example calculating partial centrality rankings with positional_dominance.

indirect_relations(g, type = "geodesic", FUN = identity, ...)

Arguments

g
igraph object. The network for which relations should be derived.

type
String giving the relation to be calculated. See Details for options.

FUN
A function that allows the transformation of relations. See Details.

...
Additional arguments passed to FUN.

Value

A matrix containing indirect relations in a network.

Details

The type parameter has the following options. 'identity' returns the adjacency matrix of the network. 'distances' returns geodesic distances between all pairs of nodes. 'dependencies' returns dyadic dependencies $$\delta(u,s) = \sum_{t \in V} \frac{\sigma(s,t|u)}{\sigma(s,t)}$$ where \(\sigma(s,t|u)\) is the number of shortest paths from s to t that include u and \(\sigma(s,t)\) is the total number of shortest (s,t)-paths. This relation is mostly used for betweenness-like centrality indices. 'walks' returns walk counts between pairs of nodes, usually they are weighted decreasingly in their lengths or other properties which can be done by adding a function in FUN. See transform_relations for options. 'resistance' returns the resistance distance between all pairs of vertices.

The function FUN is used to further specify the indirect relation. See transform_relations for predefined functions and additional help.

See also

aggregate_positions to build centrality indices, positional_dominance to derive dominance relations

Examples

require(igraph)
g <- graph.empty(n=11,directed = FALSE)
g <- add_edges(g,c(1,11,2,4,3,5,3,11,4,8,5,9,5,11,6,7,6,8,
                   6,10,6,11,7,9,7,10,7,11,8,9,8,10,9,10))
#geodesic distances
D <- indirect_relations(g,type = "geodesic")

#dyadic dependencies
D <- indirect_relations(g,type = "dependencies")

#walks attenuated exponentially by there length
W <- indirect_relations(g,type = "walks",FUN = walks_exp)

#positional dominance of a transformed relation...
D <- indirect_relations(g,type = "geodesic")

#...under total heterogeneity
positional_dominance(D, map = FALSE ,benefit = FALSE)

#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11]
#>  [1,]    0    0    1    0    1    1    1    0    0     0     1
#>  [2,]    0    0    0    1    0    0    0    1    0     0     0
#>  [3,]    0    0    0    0    1    0    0    0    0     0     1
#>  [4,]    0    0    0    0    0    0    0    0    0     0     0
#>  [5,]    0    0    0    0    0    0    0    0    0     0     0
#>  [6,]    0    0    0    0    0    0    0    0    0     0     0
#>  [7,]    0    0    0    0    0    0    0    0    0     0     0
#>  [8,]    0    0    0    0    0    0    0    0    0     0     0
#>  [9,]    0    0    0    0    0    0    0    0    0     0     0
#> [10,]    0    0    0    0    0    0    0    0    0     0     0
#> [11,]    0    0    0    0    0    0    0    0    0     0     0

#...under total homogeneity
positional_dominance(D, map = TRUE ,benefit = FALSE)

#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11]
#>  [1,]    0    0    1    1    1    1    1    1    1     1     1
#>  [2,]    1    0    1    1    1    1    1    1    1     1     1
#>  [3,]    0    0    0    0    1    1    1    1    1     1     1
#>  [4,]    0    0    0    0    1    1    1    1    1     1     1
#>  [5,]    0    0    0    0    0    1    1    1    1     0     1
#>  [6,]    0    0    0    0    0    0    0    0    0     0     0
#>  [7,]    0    0    0    0    0    1    0    1    1     0     1
#>  [8,]    0    0    0    0    0    1    0    0    1     0     0
#>  [9,]    0    0    0    0    0    1    0    1    0     0     0
#> [10,]    0    0    0    0    0    1    0    1    1     0     0
#> [11,]    0    0    0    0    0    0    0    0    0     0     0